move the dependency to xgboost 1.6.1 (#163)
* move the dependency to xgboost 1.6.1

* resolve comments
