+++ examples/XGBoost-Examples/aggregator/.gitignore
@@ -1,3 +1,4 @@
 .idea
 target
 *.iml
+*.xml
