--- examples/XGBoost-Examples/aggregator/dependency-reduced-pom.xml
@@ -1,81 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-  <parent>
-    <artifactId>sample_xgboost_examples</artifactId>
-    <groupId>com.nvidia</groupId>
-    <version>0.2.2</version>
-  </parent>
-  <modelVersion>4.0.0</modelVersion>
-  <artifactId>spark_examples_aggregator</artifactId>
-  <build>
-    <finalName>sample_xgboost_apps-${project.version}</finalName>
-    <plugins>
-      <plugin>
-        <artifactId>maven-shade-plugin</artifactId>
-        <executions>
-          <execution>
-            <phase>package</phase>
-            <goals>
-              <goal>shade</goal>
-            </goals>
-          </execution>
-        </executions>
-        <configuration>
-          <artifactSet>
-            <includes>
-              <include>com.nvidia:spark_examples_*</include>
-            </includes>
-          </artifactSet>
-        </configuration>
-      </plugin>
-    </plugins>
-  </build>
-  <dependencies>
-    <dependency>
-      <groupId>com.nvidia</groupId>
-      <artifactId>xgboost4j_3.0</artifactId>
-      <version>1.4.2-0.3.0</version>
-      <scope>compile</scope>
-    </dependency>
-    <dependency>
-      <groupId>com.nvidia</groupId>
-      <artifactId>xgboost4j-spark_3.0</artifactId>
-      <version>1.4.2-0.3.0</version>
-      <scope>compile</scope>
-    </dependency>
-    <dependency>
-      <groupId>org.scala-lang</groupId>
-      <artifactId>scala-library</artifactId>
-      <version>2.12.8</version>
-      <scope>provided</scope>
-    </dependency>
-    <dependency>
-      <groupId>org.apache.spark</groupId>
-      <artifactId>spark-sql_2.12</artifactId>
-      <version>3.0.1</version>
-      <scope>provided</scope>
-    </dependency>
-    <dependency>
-      <groupId>org.apache.spark</groupId>
-      <artifactId>spark-mllib_2.12</artifactId>
-      <version>3.0.1</version>
-      <scope>provided</scope>
-    </dependency>
-    <dependency>
-      <groupId>org.scalatest</groupId>
-      <artifactId>scalatest_2.12</artifactId>
-      <version>3.0.5</version>
-      <scope>test</scope>
-      <exclusions>
-        <exclusion>
-          <artifactId>scalactic_2.12</artifactId>
-          <groupId>org.scalactic</groupId>
-        </exclusion>
-      </exclusions>
-    </dependency>
-  </dependencies>
-  <properties>
-    <maven.compiler.target>8</maven.compiler.target>
-    <maven.compiler.source>8</maven.compiler.source>
-  </properties>
-</project>
