+++ examples/XGBoost-Examples/pom.xml
@@ -38,7 +38,7 @@
 
     <properties>
         <encoding>UTF-8</encoding>
-        <xgboost.version>1.4.2-0.3.0</xgboost.version>
+        <xgboost.version>1.6.1</xgboost.version>
         <spark.version>3.1.1</spark.version>
         <scala.version>2.12.8</scala.version>
         <scala.binary.version>2.12</scala.binary.version>
@@ -46,13 +46,13 @@
 
     <dependencies>
         <dependency>
-            <groupId>com.nvidia</groupId>
-            <artifactId>xgboost4j_3.0</artifactId>
+            <groupId>ml.dmlc</groupId>
+            <artifactId>xgboost4j-gpu_2.12</artifactId>
             <version>${xgboost.version}</version>
         </dependency>
         <dependency>
-            <groupId>com.nvidia</groupId>
-            <artifactId>xgboost4j-spark_3.0</artifactId>
+            <groupId>ml.dmlc</groupId>
+            <artifactId>xgboost4j-spark-gpu_2.12</artifactId>
             <version>${xgboost.version}</version>
         </dependency>
         <dependency>
